requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
supabase==2.3.4
httpx==0.26.0
python-jose[cryptography]==3.3.0